removed obsolete casts.

+++ lib/socks.c
@@ -513,10 +513,10 @@ CURLcode Curl_SOCKS5(const char *proxy_name,
     len = 0;
     socksreq[len++] = 1;    /* username/pw subnegotiation version */
     socksreq[len++] = (char) userlen;
-    memcpy(socksreq + len, proxy_name, (int) userlen);
+    memcpy(socksreq + len, proxy_name, userlen);
     len += (int)userlen;
     socksreq[len++] = (char) pwlen;
-    memcpy(socksreq + len, proxy_password, (int) pwlen);
+    memcpy(socksreq + len, proxy_password, pwlen);
     len += (int)pwlen;
 
     code = Curl_write_plain(conn, sock, (char *)socksreq, len, &written);
